backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
google-generativeai
email-validator
requests
